requirements.txt
Django==5.0.6
setuptools==74.0.0
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
environs==11.0.0